Split ngram_search_query.hpp and ngram_database.hpp
